removed component dismissing from settings, created useDismissRoute and added to BackgroundBlur
